refactor: get, set params 유틸 폴더로 이동
